Fix classes to standalone class files. Removed dead files.
