+++ tests/TestCase/Database/ExpressionTypeCastingTest.php
@@ -19,21 +19,11 @@ use Cake\Database\Expression\CaseExpression;
 use Cake\Database\Expression\Comparison;
 use Cake\Database\Expression\FunctionExpression;
 use Cake\Database\Expression\ValuesExpression;
-use Cake\Database\ExpressionInterface;
-use Cake\Database\Type\ExpressionTypeInterface;
-use Cake\Database\Type\StringType;
 use Cake\Database\TypeFactory;
 use Cake\Database\TypeMap;
 use Cake\Database\ValueBinder;
 use Cake\TestSuite\TestCase;
-
-class TestType extends StringType implements ExpressionTypeInterface
-{
-    public function toExpression($value): ExpressionInterface
-    {
-        return new FunctionExpression('CONCAT', [$value, ' - foo']);
-    }
-}
+use TestApp\Database\Type\TestType;
 
 /**
  * Tests for Expression objects casting values to other expressions
