+++ tests/test_app/TestApp/Config/ReadOnlyTestInstanceConfig.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace TestApp\Config;
+
+use Cake\Core\InstanceConfigTrait;
+use Exception;
+
+class ReadOnlyTestInstanceConfig
+{
+    use InstanceConfigTrait;
+
+    /**
+     * _defaultConfig
+     *
+     * Some default config
+     *
+     * @var array
+     */
+    protected $_defaultConfig = [
+        'some' => 'string',
+        'a' => [
+            'nested' => 'value',
+            'other' => 'value',
+        ],
+    ];
+
+    /**
+     * Example of how to prevent modifying config at run time
+     *
+     * @throws \Exception
+     * @param mixed $key
+     * @param mixed $value
+     * @return void
+     */
+    protected function _configWrite($key, $value = null)
+    {
+        throw new Exception('This Instance is readonly');
+    }
+}
