+++ tests/test_app/TestApp/Database/Type/IntType.php
@@ -1,0 +1,16 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\Database\Type;
+
+use Cake\Database\Type\IntegerType;
+
+/**
+ * Mock class for testing baseType inheritance
+ */
+class IntType extends IntegerType
+{
+    public function getBaseType(): string
+    {
+        return 'integer';
+    }
+}
