+++ tests/test_app/TestApp/Database/Type/UuidValue.php
@@ -1,0 +1,16 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\Database\Type;
+
+/**
+ * Value object for testing mappings.
+ */
+class UuidValue
+{
+    public $value;
+
+    public function __construct($value)
+    {
+        $this->value = $value;
+    }
+}
