+++ tests/test_app/TestApp/Log/Object/StringObject.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace TestApp\Log\Object;
+
+/**
+ * used for testing when an object is passed to a logger
+ */
+class StringObject
+{
+    /**
+     * String representation of the object
+     *
+     * @return string
+     */
+    public function __toString()
+    {
+        return 'Hey!';
+    }
+}
